# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Web Engine Platform API
  description: |
    API Documentation for the Web Engine Platform. 
    This API provides access to all features of the platform, including user management, content management, analytics, and more.
  version: 1.0.0
  contact:
    name: Web Engine Platform Support
    email: support@webengineplatform.com
    url: https://webengineplatform.com/support
  license:
    name: Proprietary
    url: https://webengineplatform.com/terms
servers:
  - url: https://api.webengineplatform.com/api/v1
    description: Production API Server
  - url: https://staging-api.webengineplatform.com/api/v1
    description: Staging API Server
  - url: https://dev-api.webengineplatform.com/api/v1
    description: Development API Server
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate a user and return access and refresh tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh tokens
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/password-reset:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send a password reset email to the user
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '204':
          description: Password reset email sent
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/password-reset/{token}:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset user password using reset token
      operationId: resetPassword
      parameters:
        - name: token
          in: path
          description: Password reset token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: Password123!
        rememberMe:
          type: boolean
          default: false
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: Password123!
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        companyName:
          type: string
          example: Acme Inc
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    PasswordResetConfirmRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          format: password
          example: NewPassword123!
    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - user
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: Access token expiry time in seconds
          example: 900
        tokenType:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        companyName:
          type: string
          example: Acme Inc
        avatar:
          type: string
          format: uri
          example: https://api.webengineplatform.com/avatars/123.jpg
        role:
          type: string
          enum: [user, admin, editor]
          example: user
        permissions:
          type: array
          items:
            type: string
          example: ["read:projects", "write:projects"]
        lastLogin:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
...
    Error:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: The request contains invalid parameters
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          example:
            - field: email
              message: Must be a valid email address
        timestamp:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        path:
          type: string
          example: /api/v1/users
        requestId:
          type: string
          example: req_12345abcde 